requirements.txt
# Power Automate Desktop Framework Requirements

# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# CLI and user interface
click>=8.1.0
tabulate>=0.9.0
rich>=13.0.0

# Logging and monitoring
loguru>=0.7.0
colorama>=0.4.6

# Database support
pyodbc>=5.0.0
pymongo>=4.6.0
redis>=5.0.0

# Excel and file operations
openpyxl>=3.1.0
pandas>=2.0.0
xlrd>=2.0.1

# Email support
secure-smtplib>=0.1.1

# Web automation utilities
selenium>=4.16.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Async support
aiofiles>=23.2.0
asyncio>=3.4.3

# Scheduling
schedule>=1.2.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0
prometheus-client>=0.19.0

# Retry mechanisms
tenacity>=8.2.0

# JSON/XML processing
xmltodict>=0.13.0
jsonschema>=4.20.0

# Encryption
cryptography>=41.0.0

# API development (optional)
fastapi>=0.109.0
uvicorn>=0.27.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
